refactor, fuzz: Split miniscript harnesses into separate files
